refactor: move kernelClient to context
The client needs to be part of a client state, right now as a hook
local state, it was being recreated innecesarily and also could not be
shared between components. Now the simple state is in redux and this
client is part of a context provider.
